saving and committing the eliteMap only after all batch iteration results have been processed, and using the fs.promises API to save the genomes and eliteMap; significantly improving the performance of this IO bottleneck!

+++ cli-app/util/qd-common.js
@@ -44,17 +44,19 @@ export function deleteAllGenomesNotInEliteMap( eliteMap, evoRunDirPath ) {
 }
 
 export function deleteAllGenomeRendersNotInEliteMap( eliteMap, evoRenderDirPath ) {
-  const eliteGenomeIds = Object.fromEntries(
-    Object.values(eliteMap.cells).filter( oneCell => oneCell.elts.length > 0).map( oneCell => [oneCell.elts[0].g, true] )
-  );
-  const wavFiles = fs.readdirSync(evoRenderDirPath).filter( file => file.endsWith('.wav') );
-  for( let wavFile of wavFiles ) {
-    // genome ID from wavFile name like "56_28_<ID>.wav" -> "<ID>"
-    const genomeId = wavFile.substring(wavFile.lastIndexOf("_")+1, wavFile.length - 4);
-    if( !eliteGenomeIds[genomeId] ) {
-      const wavFilePath = `${evoRenderDirPath}${wavFile}`;
-      console.log(`deleting genome render: ${wavFilePath}`);
-      fs.unlinkSync(wavFilePath);
+  if( fs.existsSync(evoRenderDirPath) ) {
+    const eliteGenomeIds = Object.fromEntries(
+      Object.values(eliteMap.cells).filter( oneCell => oneCell.elts.length > 0).map( oneCell => [oneCell.elts[0].g, true] )
+    );
+    const wavFiles = fs.readdirSync(evoRenderDirPath).filter( file => file.endsWith('.wav') );
+    for( let wavFile of wavFiles ) {
+      // genome ID from wavFile name like "56_28_<ID>.wav" -> "<ID>"
+      const genomeId = wavFile.substring(wavFile.lastIndexOf("_")+1, wavFile.length - 4);
+      if( !eliteGenomeIds[genomeId] ) {
+        const wavFilePath = `${evoRenderDirPath}${wavFile}`;
+        console.log(`deleting genome render: ${wavFilePath}`);
+        fs.unlinkSync(wavFilePath);
+      }
     }
   }
 }
